replace ListView with VirtualizedList
