+++ packages/interface/src/components/explorer/Explorer.tsx
@@ -18,11 +18,11 @@ import {
 } from 'phosphor-react';
 import React, { memo, useLayoutEffect, useMemo, useRef, useState } from 'react';
 
-import { FileList } from '../explorer/FileList';
 import { Inspector } from '../explorer/Inspector';
 import { WithContextMenu } from '../layout/MenuOverlay';
 import { TopBar } from '../layout/TopBar';
 import ExplorerContextMenu from './ExplorerContextMenu';
+import { VirtualizedList } from './VirtualizedList';
 
 interface Props {
 	data: ExplorerData;
@@ -39,37 +39,31 @@ export default function Explorer(props: Props) {
 		}
 	});
 
-	return (
-		<div className="relative">
-			<ExplorerContextMenu>
-				<ExplorerContent data={props.data} />
-			</ExplorerContextMenu>
-		</div>
-	);
-}
-
-const ExplorerContent = (props: Props) => {
 	const { selectedRowIndex, showInspector } = useExplorerStore((store) => ({
 		selectedRowIndex: store.selectedRowIndex,
 		showInspector: store.showInspector
 	}));
 
 	return (
-		<div className="relative flex flex-col w-full bg-gray-650">
-			<TopBar />
-			<div className="relative flex flex-row w-full max-h-full">
-				<FileList data={props.data?.items || []} context={props.data.context} />
-				{showInspector && (
-					<div className="min-w-[260px] max-w-[260px]">
-						{props.data.items[selectedRowIndex]?.id && (
-							<Inspector
-								key={props.data.items[selectedRowIndex].id}
-								data={props.data.items[selectedRowIndex]}
-							/>
+		<div className="relative">
+			<ExplorerContextMenu>
+				<div className="relative flex flex-col w-full bg-gray-650">
+					<TopBar />
+					<div className="relative flex flex-row w-full max-h-full">
+						<VirtualizedList data={props.data?.items || []} context={props.data.context} />
+						{showInspector && (
+							<div className="min-w-[260px] max-w-[260px]">
+								{props.data.items[selectedRowIndex]?.id && (
+									<Inspector
+										key={props.data.items[selectedRowIndex].id}
+										data={props.data.items[selectedRowIndex]}
+									/>
+								)}
+							</div>
 						)}
 					</div>
-				)}
-			</div>
+				</div>
+			</ExplorerContextMenu>
 		</div>
 	);
-};
+}
